requirements.txt
# Python 3.10+. Use: python -m venv .venv && source .venv/bin/activate && pip install -r requirements.txt

# Agents & API
google-genai
python-dotenv
pydantic
fastapi
uvicorn

# Notebooks (render + kernel)
markdown
ipython
ipykernel

# Price model (scikit-learn fixado: artefato treinado com 1.4.2)
pandas
scikit-learn==1.4.2
joblib
